% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stress_test_multivariate_t_distribution.R
\name{stress_test_multivariate_t_distribution}
\alias{stress_test_multivariate_t_distribution}
\title{Stress-Test the Correlation Structure of a Multivariate t Distribution}
\usage{
stress_test_multivariate_t_distribution(
  .num_assets = 5,
  .sample_size = 200,
  .simulations = 500,
  .nu = 4
)
}
\arguments{
\item{.num_assets}{A \code{numeric} scalar. It's the dimension of the of simulated
market.}

\item{.sample_size}{A \code{numeric} scalar. It's the number of realizations of
each series.}

\item{.simulations}{A \code{numeric} scalar. The number of simulations to be
executed for each level of correlation.}

\item{.nu}{A \code{numeric} scalar. The degrees of freedom.}
}
\value{
A \code{ggplot} object.
}
\description{
This function stresses the first two moments of a normal t distribution
in a variety of correlation environments.
}
\examples{
#
}
