% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stress_test_univariate_t_distribution.R
\name{stress_test_univariate_t_distribution}
\alias{stress_test_univariate_t_distribution}
\alias{stress_test_univariate_t_distribution.default}
\alias{stress_test_univariate_t_distribution.tbl}
\alias{stress_test_univariate_t_distribution.xts}
\alias{stress_test_univariate_t_distribution.matrix}
\alias{stress_test_univariate_t_distribution.numeric}
\title{Stress-Test the Student-t Distribution}
\usage{
stress_test_univariate_t_distribution(
  .invariant,
  .what = c("mu", "sigma", "nu", "skew"),
  .simulations = 300,
  .symmetric = FALSE
)

\method{stress_test_univariate_t_distribution}{default}(
  .invariant,
  .what = c("mu", "sigma", "nu", "skew"),
  .simulations = 300,
  .symmetric = FALSE
)

\method{stress_test_univariate_t_distribution}{tbl}(
  .invariant,
  .what = c("mu", "sigma", "nu", "skew"),
  .simulations = 300,
  .symmetric = FALSE
)

\method{stress_test_univariate_t_distribution}{xts}(
  .invariant,
  .what = c("mu", "sigma", "nu", "skew"),
  .simulations = 300,
  .symmetric = FALSE
)

\method{stress_test_univariate_t_distribution}{matrix}(
  .invariant,
  .what = c("mu", "sigma", "nu", "skew"),
  .simulations = 300,
  .symmetric = FALSE
)

\method{stress_test_univariate_t_distribution}{numeric}(
  .invariant,
  .what = c("mu", "sigma", "nu", "skew"),
  .simulations = 300,
  .symmetric = FALSE
)
}
\arguments{
\item{.invariant}{An univariate data structure.}

\item{.what}{A \code{character} indicating which parameter should be stressed.}

\item{.simulations}{A \code{numeric} scalar. The number of simulations to be
executed for each level of correlation.}

\item{.symmetric}{A \code{character} flag. Should the distributions be symmetric? Default to FALSE.}
}
\value{
A \code{ggplot} object.
}
\description{
This function implements a stress-test in the NIG by stretching one the parameters
(mean, sigma or skewness) by 20 percent in both directions (up and down). Then, analyses
the impact that the shaked parameter causes in all the other parameters.
}
\examples{
#
}
